requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
urllib3==2.2.1
charset-normalizer==3.3.2
gunicorn==21.2.0